news-extractor-backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Devin AI <158243242+devin-ai-integration[bot]@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.135.1"}
psycopg = {extras = ["binary"], version = "^3.3.3"}
httpx = "^0.28.1"
beautifulsoup4 = "^4.14.3"
python-dotenv = "^1.2.2"
google-genai = "^1.65.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
